backend/requirements.txt
fastapi
uvicorn
pydantic==2.9.2
python-dotenv
h3
requests==2.32.3
rasterio==1.3.10
numpy==1.26.4
pandas==2.2.2 
pyarrow==15.0.2 
tqdm==4.66.4
openai